chore: View 클래스 변수 타입 변경
수익률 변수 float을 double로 변경

+++ src/main/java/lotto/View.java
@@ -6,7 +6,7 @@ import static camp.nextstep.edu.missionutils.Console.readLine;
 
 public class View {
     private Map<Integer, Integer> userWinningRanks;
-    private float profitSummary;
+    private double profitSummary;
     private int lottoPurchaseNum;
     private String winningNumber;
     private String bounsNumber;
